agentic_retrieval/pyproject.toml
[project]
name = "browsecomp-plus"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.9.0",
    "anthropic>=0.58.2",
    "datasets>=4.0.0",
    "deepspeed>=0.17.2",
    "faiss-cpu>=1.11.0.post1",
    "fastmcp==2.9.2",
    "google-genai>=1.27.0",
    "openai>=1.97.0",
    "peft>=0.16.0",
    "pyngrok>=7.2.12",
    "pyserini>=1.2.0",
    "python-dotenv>=1.1.1",
    "qwen-agent[gui,mcp,rag]==0.0.27",
    "qwen-omni-utils==0.0.8",
    "rich>=14.0.0",
    "tevatron",
    "torchvision",
    "tqdm>=4.67.1",
    "sentence-transformers>=2.0.0",
    "transformers>=4.53.2",
    "vllm>=0.9.0",
    "gritlm>=1.0.2",
    "gensim>=4.3.3,<5",
]

[tool.uv.sources]
tevatron = { git = "https://github.com/texttron/tevatron.git", rev = "main" }
